fixed PC offset

+++ include/vm.h
@@ -18,7 +18,7 @@ typedef struct {
 	vm_stack stack;
 
 	uint32_t pc;
-	uint32_t hsize;
+	size_t hsize;
 
 	uint32_t* registers;
 } vm_runtime;
